Update main notebook to incorporate graph-based MIDI approach
The main Jupyter notebook has been updated to use a graph-based approach for MIDI file handling. The previous method of MIDI handling has been replaced with functions to convert, train and visualise MIDI files as graphs. This new approach allows for a more efficient and detailed analysis of MIDI files.