Add MIDI files graph conversion and visualization functionalities
A graph.py module is added that enables conversion of MIDI files into graph representation,the nodes for notes, chords, and instrument changes are properly identified and added. Meanwhile, a train.py module was implemented to train the graph with node2vec. Lastly, introduced a function in the visualise.py module that uses matplotlib to visually render the graph.